# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: VPN Pulse API
  version: 1.5.1
  description: Closed Telegram Mini App and probe API. Unknown values are null, never zero.
servers:
  - url: /api/v1
...
paths:
...
  /servers/{serverId}:
    get:
      tags: [Status]
      summary: Member-safe server details
      operationId: getServer
      parameters:
        - $ref: "#/components/parameters/AcceptLanguage"
      security: [{ webSession: [] }]
      parameters:
        - $ref: "#/components/parameters/ServerId"
      responses:
        "200":
          description: Server details
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ServerDetail" }
        "404": { $ref: "#/components/responses/Problem" }
...
  /admin/servers/{serverId}:
    get:
      tags: [Admin]
      summary: Admin-only diagnostic projection
      operationId: getAdminServer
      parameters:
        - $ref: "#/components/parameters/AcceptLanguage"
      security: [{ webSession: [] }]
      parameters:
        - $ref: "#/components/parameters/ServerId"
      responses:
        "200":
          description: Diagnostics
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminServerDetail" }
        "403": { $ref: "#/components/responses/Problem" }
...
components:
  securitySchemes:
    webSession:
      type: apiKey
      in: cookie
      name: vpnpulse_session
    probeToken:
      type: http
      scheme: bearer
      bearerFormat: opaque
  parameters:
    AcceptLanguage:
      name: Accept-Language
      in: header
      required: false
      description: >-
        language of server names, the administrator note and doctor hints: the first supported tag
        (ru or en) wins, otherwise the installation default. Everything else is an i18n key the
        client renders itself.
      schema: { type: string, example: "ru-RU,ru;q=0.9,en;q=0.5" }
    ServerId:
      name: serverId
      in: path
      required: true
      schema: { $ref: "#/components/schemas/PublicId" }
  responses:
    Problem:
      description: RFC 9457 problem details
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
  schemas:
    PublicId:
      type: string
      pattern: "^[a-z][a-z0-9-]{1,63}$"
    State:
      type: string
      enum: [operational, degraded, unavailable, unknown]
    Freshness:
      type: object
      additionalProperties: false
      required: [observed_at, fresh_until, is_stale]
      properties:
        observed_at: { type: [string, "null"], format: date-time }
        fresh_until: { type: [string, "null"], format: date-time }
        is_stale: { type: boolean }
    EvidenceSummary:
      type: object
      additionalProperties: false
      required: [source, state, freshness]
      properties:
        source: { type: string, enum: [collector, pc, mobile, abroad, human_activity] }
        state: { $ref: "#/components/schemas/State" }
        freshness: { $ref: "#/components/schemas/Freshness" }
        reason_code: { type: [string, "null"], maxLength: 80 }
        via_server_id:
          description: for `abroad` evidence — the public id of the server that performed the cross-check
          anyOf: [{ $ref: "#/components/schemas/PublicId" }, { type: "null" }]
    ServerCard:
      type: object
      required: [id, name, country_code, state, freshness, recommended, sources, uptime_24h]
      properties:
        id: { $ref: "#/components/schemas/PublicId" }
        name: { type: string, minLength: 1, maxLength: 80 }
        country_code: { type: string, pattern: "^[A-Z]{2}$" }
        state: { $ref: "#/components/schemas/State" }
        freshness: { $ref: "#/components/schemas/Freshness" }
        recommended: { type: boolean }
        uptime_24h: { type: [number, "null"], minimum: 0, maximum: 1 }
        coverage_24h: { type: [number, "null"], minimum: 0, maximum: 1 }
        sources:
          type: array
          items: { $ref: "#/components/schemas/EvidenceSummary" }
...
    ProtocolSummary:
      type: object
      additionalProperties: false
      required: [kind, state, coverage, connections]
      properties:
        kind: { type: string, enum: [amneziawg, xray_reality] }
        state: { $ref: "#/components/schemas/State" }
        coverage: { type: number, minimum: 0, maximum: 1 }
        connections: { type: [integer, "null"], minimum: 0 }
    ServerDetail:
      allOf:
        - $ref: "#/components/schemas/ServerCard"
        - type: object
          required: [protocols, checks]
          properties:
            protocols:
              type: array
              items: { $ref: "#/components/schemas/ProtocolSummary" }
            checks:
              description: only sources listed in /status.sources; empty when no probe has reported yet
              type: array
              items: { $ref: "#/components/schemas/EvidenceSummary" }
            uptime_7d: { type: [number, "null"], minimum: 0, maximum: 1 }
            coverage_7d: { type: [number, "null"], minimum: 0, maximum: 1 }
            components:
              description: member-safe software list for the "Software" group (no hosts, ports or paths)
              type: array
              maxItems: 12
              items: { $ref: "#/components/schemas/SoftwareComponent" }
            resources: { anyOf: [{ $ref: "#/components/schemas/ResourceSnapshot" }, { type: "null" }] }
            profiles: { anyOf: [{ $ref: "#/components/schemas/ProfileStats" }, { type: "null" }] }
            service_checks:
              description: binary collector checks for the "Server" group
              type: array
              maxItems: 8
              items: { $ref: "#/components/schemas/ServiceCheck" }
    SoftwareComponent:
      type: object
      additionalProperties: false
      required: [id, name]
      properties:
        id: { type: string, enum: [amneziawg, xray, hiddify, ddns, tuning, other] }
        name: { type: string, maxLength: 40 }
        version: { type: [string, "null"], maxLength: 32 }
        note_key: { type: [string, "null"], pattern: "^server[.]software[.]" }
        note: { type: [string, "null"], maxLength: 60, description: free member-safe detail such as the transport name; shown when note_key is null }
    ResourceSnapshot:
      description: aggregate load; null means the collector could not measure it
      type: object
      additionalProperties: false
      required: [traffic_mbps, cpu_percent, memory_percent]
      properties:
        traffic_mbps: { type: [number, "null"], minimum: 0 }
        cpu_percent: { type: [number, "null"], minimum: 0, maximum: 100 }
        memory_percent: { type: [number, "null"], minimum: 0, maximum: 100 }
        swap_percent: { type: [number, "null"], minimum: 0, maximum: 100 }
        peak_connections_24h: { type: [integer, "null"], minimum: 0 }
    ProfileStats:
      description: aggregate VPN profile counts, never per-person data
      type: object
      additionalProperties: false
      required: [issued, ever_connected, active_24h]
      properties:
        issued: { type: [integer, "null"], minimum: 0 }
        ever_connected: { type: [integer, "null"], minimum: 0 }
        active_24h: { type: [integer, "null"], minimum: 0 }
        last_connection_at: { type: [string, "null"], format: date-time, description: last handshake seen by the collector; null when never or unknown }
    ServiceCheck:
      type: object
      additionalProperties: false
      required: [check, ok]
      properties:
        check: { type: string, enum: [ssh, engine, port, dns, ddns] }
        ok: { type: [boolean, "null"] }
...
    AdminServerDetail:
      unevaluatedProperties: false
      allOf:
        - $ref: "#/components/schemas/ServerDetail"
        - type: object
          required: [attention, diagnostics]
          properties:
            attention: { type: array, items: { type: string, maxLength: 80 } }
            attention_items:
              description: structured attention list for the admin dashboard (severity-ordered)
              type: array
              maxItems: 50
              items: { $ref: "#/components/schemas/AttentionItem" }
            diagnostics: { type: object, additionalProperties: true }
...
    AttentionItem:
      type: object
      additionalProperties: false
      required: [severity, code, message]
      properties:
        severity: { type: string, enum: [high, medium, low] }
        code: { type: string, pattern: "^[A-Z][A-Z0-9_]+$" }
        message: { type: string, maxLength: 240 }
        server_id: { anyOf: [{ $ref: "#/components/schemas/PublicId" }, { type: "null" }] }
...
    Problem:
      type: object
      additionalProperties: false
      required: [type, title, status, code, trace_id]
      properties:
        type: { type: string, format: uri-reference }
        title: { type: string }
        status: { type: integer, minimum: 400, maximum: 599 }
        code: { type: string, pattern: "^[A-Z][A-Z0-9_]+$" }
        trace_id: { type: string, minLength: 8, maxLength: 64 }